Add video replacement feature.
This will be important if we want to transition from the player to the
list. When the parent HTML changes, the virtual DOM thinks that it needs
to make a new video player. The Replace message tells JavaScript that,
on the next animation frame, it should replace the new video element
with the old one. This process is completely invisible except rarely a
very brief pause in the video. This replacement technique should work
regardless of where the video elements are or how they are styled.